requirements.txt
antlr4-python3-runtime==4.9.3
argcomplete==3.6.0
asteroid-filterbanks==0.4.0
audioread==3.0.1
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ci-sdr==0.0.2
click==8.1.8
ConfigArgParse==1.7
contourpy==1.3.1
ctc_segmentation==1.7.4
cycler==0.12.1
Cython==3.0.12
decorator==5.2.1
Distance==0.1.3
editdistance==0.8.1
einops==0.8.1
espnet==202412
espnet-tts-frontend==0.0.3
fast-bss-eval==0.1.3
filelock==3.18.0
fonttools==4.56.0
fsspec==2025.3.0
g2p-en==2.1.0
gdown==5.2.0
h5py==3.13.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
importlib-metadata==4.13.0
inflect==7.5.0
jaconv==0.4.0
jamo==0.4.1
Jinja2==3.1.6
joblib==1.4.2
kaiser==0.1
kaldiio==2.18.1
kiwisolver==1.4.8
lazy_loader==0.4
librosa==0.9.2
llvmlite==0.43.0
MarkupSafe==3.0.2
matplotlib==3.10.1
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
networkx==3.4.2
nltk==3.9.1
numba==0.60.0
numpy==1.23.5
omegaconf==2.3.0
opt_einsum==3.4.0
packaging==24.2
parallel-wavegan==0.6.1
pillow==11.1.0
platformdirs==4.3.7
pooch==1.8.2
protobuf==6.30.1
pycparser==2.22
pyparsing==3.2.1
pypinyin==0.44.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pyworld==0.3.5
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
resampy==0.4.3
scikit-learn==1.6.1
scipy==1.10.0
sentencepiece==0.1.97
six==1.17.0
soundfile==0.13.1
soupsieve==2.6
soxr==0.5.0.post1
sympy==1.13.1
tensorboardX==2.6.2.2
threadpoolctl==3.6.0
tomlkit==0.13.2
torch==2.6.0
torch-complex==0.4.4
torchaudio==2.6.0
tqdm==4.67.1
typeguard==4.4.2
typing_extensions==4.12.2
Unidecode==1.3.8
urllib3==2.3.0
xmltodict==0.14.2
yq==3.4.3
zipp==3.21.0
